fix: 직업 경험치 획득 표기를 실제 수치로 변경
- 경험치 수집기를 정수에서 실수 기반으로 변경
- 액션바 표기를 실제 획득량 기준 소수점 첫째 자리로 표시
- 레벨업 차감이 획득 메시지를 가리던 문제 함께 해결

+++ JobsPlusRemastered-26.2/common/src/main/java/com/daqem/jobsplus/player/job/exp/ExpCollector.java
@@ -5,21 +5,21 @@ import com.daqem.jobsplus.JobsPlus;
 public class ExpCollector
 {
 
-    private int jobExp = 0;
+    private double jobExp = 0.0D;
 
-    public void addExp(int exp)
+    public void addExp(double exp)
     {
         JobsPlus.debug("Adding {} experience to the job.", exp);
         jobExp += exp;
     }
 
-    public int getExp()
+    public double getExp()
     {
         return jobExp;
     }
 
     public void clear()
     {
-        jobExp = 0;
+        jobExp = 0.0D;
     }
 }
